Made test for cancelling a pending job a integration test as it uses Octopus and runs shell commands.
The test to verify the http client can handle MAC authentication has been refactored.
Instead of using an external http server to test against in now uses an internal one.
The internal http server is part of the apache http client tests.
The integration config file is not needed anymore.